Change data type of iterator

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -12,7 +12,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int sub;
+	unsigned int sub;
 
 	if (!array || !action)
 		return;
